Test(EventResultTest) : EventApplierTest삭제 및 EventResult 도메인에 대한 테스트 코드 작성
- EventApplierTest삭제
  - EventApplier 도메인의 이름이 EventResult로 변경됨
  - 이벤트 적용에 대한 역할은 SaleEvent, GiftEvent로 각각 분배함
